persist: rename mod r#impl to internal
Using raw identifiers for module names is allowed by rust but it seems
break tooling at times, so it was probably too clever.
