+++ src/persist-client/build.rs
@@ -10,6 +10,6 @@
 fn main() {
     prost_build::Config::new()
         .type_attribute(".", "#[derive(serde::Serialize)]")
-        .compile_protos(&["persist-client/src/impl/state.proto"], &[".."])
+        .compile_protos(&["persist-client/src/internal/state.proto"], &[".."])
         .unwrap();
 }
