+++ src/persist-client/src/internal/compact.rs
@@ -26,9 +26,9 @@ use tracing::{debug_span, warn, Instrument, Span};
 
 use crate::async_runtime::CpuHeavyRuntime;
 use crate::batch::BatchParts;
-use crate::r#impl::machine::{retry_external, Machine};
-use crate::r#impl::state::HollowBatch;
-use crate::r#impl::trace::FueledMergeRes;
+use crate::internal::machine::{retry_external, Machine};
+use crate::internal::state::HollowBatch;
+use crate::internal::trace::FueledMergeRes;
 use crate::read::fetch_batch_part;
 use crate::{Metrics, PersistConfig, ShardId, WriterId};
