+++ src/persist-client/src/internal/trace.rs
@@ -58,7 +58,7 @@ use timely::PartialOrder;
 #[allow(unused_imports)] // False positive.
 use mz_ore::fmt::FormatBuffer;
 
-use crate::r#impl::state::HollowBatch;
+use crate::internal::state::HollowBatch;
 
 #[derive(Debug, Clone, PartialEq)]
 pub struct FueledMergeReq<T> {
@@ -1051,7 +1051,7 @@ impl<T: Timestamp + Lattice> MergeVariant<T> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::r#impl::paths::PartialBlobKey;
+    use crate::internal::paths::PartialBlobKey;
 
     #[test]
     fn trace_datadriven() {
