+++ src/persist-client/src/read.rs
@@ -32,11 +32,11 @@ use mz_persist::location::Blob;
 use mz_persist_types::{Codec, Codec64};
 
 use crate::error::InvalidUsage;
-use crate::r#impl::encoding::SerdeReaderEnrichedHollowBatch;
-use crate::r#impl::machine::{retry_external, Machine};
-use crate::r#impl::metrics::Metrics;
-use crate::r#impl::paths::PartialBlobKey;
-use crate::r#impl::state::{HollowBatch, Since};
+use crate::internal::encoding::SerdeReaderEnrichedHollowBatch;
+use crate::internal::machine::{retry_external, Machine};
+use crate::internal::metrics::Metrics;
+use crate::internal::paths::PartialBlobKey;
+use crate::internal::state::{HollowBatch, Since};
 use crate::{GarbageCollector, PersistConfig, ShardId};
 
 /// An opaque identifier for a reader of a persist durable TVC (aka shard).
@@ -946,7 +946,7 @@ mod tests {
     use mz_persist::unreliable::{UnreliableConsensus, UnreliableHandle};
     use timely::ExchangeData;
 
-    use crate::r#impl::metrics::Metrics;
+    use crate::internal::metrics::Metrics;
     use crate::tests::all_ok;
     use crate::{PersistClient, PersistConfig};
